pyproject.toml
[project]
name = "agent-development-kit-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk[database]>=1.31.0",
    "google-generativeai>=0.8.6",
    "litellm>=1.83.9",
    "psutil>=7.2.2",
    "python-dotenv>=1.0.1",
    "yfinance>=1.3.0",
]
